+++ spec/yard/lint/validators/tags/informal_notation_spec.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+RSpec.describe Yard::Lint::Validators::Tags::InformalNotation do
+  it 'has all required components' do
+    expect(defined?(described_class)).to be_truthy
+    expect(defined?(described_class::Config)).to be_truthy
+    expect(defined?(described_class::Parser)).to be_truthy
+    expect(defined?(described_class::Validator)).to be_truthy
+    expect(defined?(described_class::Result)).to be_truthy
+    expect(defined?(described_class::MessagesBuilder)).to be_truthy
+  end
+end